[ADD]: prevent delete if have navigation

+++ PRN222.Assignment.FPTURoomBooking.Services/Services/CampusService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Linq.Expressions;
 using Mapster;
+using Microsoft.EntityFrameworkCore;
 using PRN222.Assignment.FPTURoomBooking.Repositories.Models;
 using PRN222.Assignment.FPTURoomBooking.Repositories.UnitOfWork;
 using PRN222.Assignment.FPTURoomBooking.Services.Models.Campus;
@@ -27,11 +28,19 @@ namespace PRN222.Assignment.FPTURoomBooking.Services.Services
 
         public async Task<Result> DeleteAsync(Guid id)
         {
-            var entity = await _unitOfWork.CampusRepository.GetByIdAsync(id);
+            var entity = await _unitOfWork.CampusRepository.GetQueryable()
+                .Include(x => x.Departments)
+                .FirstOrDefaultAsync(x => x.Id == id);
             if (entity == null)
             {
                 return Result.Failure("Campus not found");
             }
+
+            if (entity.Departments.Any())
+            {
+                return Result.Failure("Cannot delete campus with associated departments");
+            }
+
             _unitOfWork.CampusRepository.Remove(entity);
             await _unitOfWork.SaveChangesAsync();
             return Result.Success();
@@ -44,6 +53,7 @@ namespace PRN222.Assignment.FPTURoomBooking.Services.Services
             {
                 return Result<CampusModel>.Failure("Campus not found");
             }
+
             return entity.Adapt<CampusModel>();
         }
